+++ server/repository/space-repository.go
@@ -29,14 +29,16 @@ type Space struct {
 }
 
 type SpaceAvailabilityBookingEntry struct {
-	BookingID   string
-	RecurringID string
-	UserID      string
-	UserEmail   string
-	Enter       time.Time
-	Leave       time.Time
-	Subject     string
-	Approved    bool
+	BookingID     string
+	RecurringID   string
+	UserID        string
+	UserEmail     string
+	UserFirstname string
+	UserLastname  string
+	Enter         time.Time
+	Leave         time.Time
+	Subject       string
+	Approved      bool
 }
 
 type SpaceAvailability struct {
@@ -159,7 +161,7 @@ func (r *SpaceRepository) GetAllInTime(locationID string, enter, leave time.Time
 		")"
 	rows, err := GetDatabase().DB().Query("SELECT id, location_id, name, x, y, width, height, rotation, require_subject, enabled, kiosk_enabled, shape, "+
 		"NOT EXISTS(SELECT id FROM bookings WHERE "+subQueryWhere+"), "+
-		"ARRAY(SELECT CONCAT(users.id, '@@@', users.email, '@@@', bookings.enter_time, '@@@', bookings.leave_time, '@@@', bookings.id, '@@@', bookings.subject, '@@@', bookings.recurring_id, '@@@', bookings.approved::text) FROM bookings INNER JOIN users ON users.id = bookings.user_id WHERE "+subQueryWhere+" ORDER BY bookings.enter_time ASC) "+
+		"ARRAY(SELECT CONCAT(users.id, '@@@', users.email, '@@@', bookings.enter_time, '@@@', bookings.leave_time, '@@@', bookings.id, '@@@', bookings.subject, '@@@', bookings.recurring_id, '@@@', bookings.approved::text, '@@@', COALESCE(users.firstname, ''), '@@@', COALESCE(users.lastname, '')) FROM bookings INNER JOIN users ON users.id = bookings.user_id WHERE "+subQueryWhere+" ORDER BY bookings.enter_time ASC) "+
 		"FROM spaces "+
 		"WHERE location_id = $3 "+
 		"ORDER BY name", enter, leave, locationID)
@@ -177,14 +179,16 @@ func (r *SpaceRepository) GetAllInTime(locationID string, enter, leave time.Time
 			enter, _ := time.Parse(timeFormat, tokens[2])
 			leave, _ := time.Parse(timeFormat, tokens[3])
 			entry := &SpaceAvailabilityBookingEntry{
-				BookingID:   tokens[4],
-				UserID:      tokens[0],
-				UserEmail:   tokens[1],
-				Enter:       enter,
-				Leave:       leave,
-				Subject:     tokens[5],
-				RecurringID: tokens[6],
-				Approved:    tokens[7] == "true",
+				BookingID:     tokens[4],
+				UserID:        tokens[0],
+				UserEmail:     tokens[1],
+				UserFirstname: tokens[8],
+				UserLastname:  tokens[9],
+				Enter:         enter,
+				Leave:         leave,
+				Subject:       tokens[5],
+				RecurringID:   tokens[6],
+				Approved:      tokens[7] == "true",
 			}
 			e.Bookings = append(e.Bookings, entry)
 		}