+++ ui/src/pages/admin/settings/auth-providers/[id].tsx
@@ -165,8 +165,10 @@ class EditAuthProvider extends React.Component<Props, State> {
         code = e.appErrorCode;
       }
       this.setState({
-        error: code != 0,
-        errorText: code ? ErrorText.getTextForAppCode(code, this.props.t) : "",
+        error: true,
+        errorText: code
+          ? ErrorText.getTextForAppCode(code, this.props.t)
+          : this.props.t("errorSave"),
       });
     }
   };
@@ -281,11 +283,7 @@ class EditAuthProvider extends React.Component<Props, State> {
     if (this.state.saved) {
       hint = <Alert variant="success">{this.props.t("entryUpdated")}</Alert>;
     } else if (this.state.error) {
-      hint = (
-        <Alert variant="danger">
-          {this.state.errorText ?? this.props.t("errorSave")}
-        </Alert>
-      );
+      hint = <Alert variant="danger">{this.state.errorText}</Alert>;
     }
 
     let callbackUrlInfo = <></>;