+++ ui/src/util/BrowserUtil.ts
@@ -1,9 +1,11 @@
 export default class BrowserUtil {
   static LOCAL_STORAGE_KEY_SEARCH_VIEW = "searchListView";
   static LOCAL_STORAGE_KEY_MY_BOOKINGS_VIEW = "myBookingsListView";
+  static LOCAL_STORAGE_KEY_SEARCH_BOOKER_NAMES = "searchBookerNames";
 
   static tryLocalStorageSetItem(key: string, value: string): boolean {
-    if (window === undefined || window.localStorage === undefined) return false;
+    if (typeof window === "undefined" || window.localStorage === undefined)
+      return false;
     try {
       window.localStorage.setItem(key, value);
     } catch {
@@ -13,7 +15,7 @@ export default class BrowserUtil {
   }
 
   static tryLocalStorageGetItem(key: string, defaultValue: any): any {
-    if (window === undefined || window.localStorage === undefined)
+    if (typeof window === "undefined" || window.localStorage === undefined)
       return defaultValue;
     try {
       return window.localStorage.getItem(key) ?? defaultValue;